Feature uuid pk

See merge request cynergi2019/cynergi-middleware!287

+++ src/main/kotlin/com/cynergisuite/middleware/accounting/account/payable/payment/AccountPayablePaymentStatusType.kt
@@ -3,13 +3,13 @@ package com.cynergisuite.middleware.accounting.account.payable.payment
 import com.cynergisuite.domain.TypeDomainEntity
 
 data class AccountPayablePaymentStatusType(
-   val id: Long,
+   val id: Int,
    val value: String,
    val description: String,
    val localizationCode: String
 ) : TypeDomainEntity<AccountPayablePaymentStatusType> {
 
-   override fun myId(): Long = id
+   override fun myId(): Int = id
    override fun myValue(): String = value
    override fun myDescription(): String = description
    override fun myLocalizationCode(): String = localizationCode
